+++ data/src/main/java/com/london/data/datasource/remote/details/moviedetails/MovieDetailsRemote.kt
@@ -1,7 +1,7 @@
 package com.london.data.datasource.remote.details.moviedetails
 
-import com.london.data.datasource.remote.details.moviedetails.model.MovieDetailsResponse
 import com.london.data.datasource.remote.details.moviedetails.model.moviecast.MovieCastResponse
+import com.london.data.datasource.remote.details.moviedetails.model.moviedetails.MovieDetailsResponse
 import com.london.data.datasource.remote.details.moviedetails.model.movieimages.MovieImagesResponse
 import com.london.data.datasource.remote.details.moviedetails.model.similarmovies.SimilarMoviesResponse
