+++ data/src/main/java/com/london/data/mapper/moviedetails/GenreMapper.kt
@@ -1,6 +1,6 @@
 package com.london.data.mapper.moviedetails
 
-import com.london.data.datasource.remote.details.moviedetails.model.GenreRemote
+import com.london.data.datasource.remote.details.moviedetails.model.moviedetails.GenreRemote
 import com.london.domain.entity.moviedatails.Genre
 
 fun GenreRemote.toGenre(): Genre {
